Global update. Fixed swipe (adb client), swipeInRect (adb auto), added workable and tested login script (with new achitecture), new syscall - processService (in core) and added debug mode to Logger

+++ services/adb_manager/manager.py
@@ -41,7 +41,8 @@ class AdbClient:
             )
         )
 
-    def swipe(self, dotStart: Dot, dotFinish: Dot, timeK: float=0.02, swipeTime: Optional[float]=None) -> bool:
+    def swipe(self, dotStart: Dot, dotFinish: Dot, timeK: float=0.02, swipeTime: Optional[int]=None) -> bool:
+        # swipe time in seconds (CANNOT BE FLOAT (IDK WHY))
         swipeTime = swipeTime * 1000 if swipeTime else int((dotFinish - dotStart) * timeK)
         return bool(
             self.sendAdbCommand(
